+++ Jetson_Nano/face_mask_detection_model/variables/variables.index
